[IMP] coupon,gift_card,loyalty/saas~15.3: refactor loyalties completely
Complete migration script for `coupon`, `gift_card`, `pos_loyalty` into
the new system with the `loyalty` module as base.

The script will try to create coherent data and keep history for every
promotion, coupon or gift card ever used.

Due to `pos_loyalty` sharing the same model/table names as our new
module `loyalty` while not sharing the same fields its tables are
renamed to be used further down the upgrade scripts.

Relates to https://github.com/odoo/odoo/pull/82250
Relates to https://github.com/odoo/enterprise/pull/23256

TaskId-2675382

closes odoo/upgrade#3234

Related: odoo/odoo#82250
Related: odoo/enterprise#23256

+++ migrations/pos_loyalty/saas~15.1.1.0/pre-migrate.py
@@ -1,6 +1,11 @@
 # -*- coding: utf-8 -*-
 
+from odoo.upgrade import util
+
 
 def migrate(cr, version):
-    cr.execute("DELETE FROM loyalty_rule WHERE loyalty_program_id IS NULL")
-    cr.execute("DELETE FROM loyalty_reward WHERE loyalty_program_id IS NULL")
+    # Since pos_loyalty may now depend on loyalty(starting saas~15.3), loyalty's migration might have already
+    #  started, in that case it was already handled by loyalty's pre-migrate.
+    if not util.table_exists(cr, "pos_loyalty_program"):
+        cr.execute("DELETE FROM loyalty_rule WHERE loyalty_program_id IS NULL")
+        cr.execute("DELETE FROM loyalty_reward WHERE loyalty_program_id IS NULL")
